requirements.txt
et-xmlfile==1.1.0
numpy==1.25.1
openpyxl==3.1.2
pandas==2.0.3
psycopg2-binary==2.9.6
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
six==1.16.0
tzdata==2023.3
